Moved all files inside Class-based folder
